+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.9.1 (2019-01-05)
+
+### Added
+
+- Missing phpdoc typehints for `VideoInfoInterface`
+
 ## 0.9.0 (2019-01-05)
 
 ### Added